# Use sync/atomic.Value to store certificate and avoid sync.RWMutex.